Change PendingCipherChangeType storage from Int16 to String and offlinePasswordChangeCount from Int16 to Int
String-backed enum storage prevents silent data corruption from case
reordering, is self-documenting when inspecting the database, and aligns
with how the rest of the project serializes enums via Codable JSON blobs.
Int (Integer 64) for the count removes the artificial Int16 constraint.

+++ BitwardenShared/Core/Vault/Services/OfflineSyncResolverTests.swift
@@ -55,7 +55,7 @@ class OfflineSyncResolverTests: BitwardenTestCase {
         changeType: PendingCipherChangeType,
         cipherData: Data? = nil,
         originalRevisionDate: Date? = nil,
-        offlinePasswordChangeCount: Int16 = 0
+        offlinePasswordChangeCount: Int = 0
     ) async throws {
         let dataStore = DataStore(errorReporter: MockErrorReporter(), storeType: .memory)
         try await dataStore.upsertPendingChange(
